+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## v0.4.4 - September 20, 2020
 * Support loading a decrypted titlekey for `CDNReader`
 * Remove unused `sections` attribute from `CDNReader`
 * Add `available_sections` to `CDNReader` to provide a list of sections available in the title